Setup CircleCI (#22)
* Add CircleCI configuration.
Configurations includes caching of pods for faster updates, building and testing code, and uploading coverage report to CodeCov.

* Adding Gemfile and Gemfile.lock used in CircleCI.
Gemfile specifies a specific CocoaPods version while the Gemfile.lock is used for caching on the CI system.
Used in CircleCI to cache CocoaPods to speed up build times.

* Remove GitHub Actions configuration

* CircleCI config pre-start simulator.
Pre-start the iPhone simulator in the CircleCI configuration to prevent delays and timeouts while testing. This is a recommended step by CircleCI.

* Fix indentation

* Fix indentation again to pass config validation.

* Reference environment variable correctly.

* Pre-start iPhone 8 (13.0) simulator.
In the macOS container in CircleCI, Xcode 11.0.0 only supports iOS runtimes 11.4, 12.2, and 13.0. Changing pre-start simulator task to use one of the available runtimes.