[AMSDK-9772][CI] Build and save lib (#29)
* [amsdk-9772] Set Xcode config archs to all supported archs.
Set architectures and valid_architectures for the project to armv7, armv7s, arm64, i386, and x86_64. Set architectures and valid_architectures to '' for each target.

* [amsdk-9772] copy final library to 'out' folder.
Copying the final lib from the bin to out folder makes it easier to simply archive the 'out' folder.

* [amsdk-9772] Add 'deploy' job to CircleCI to archive lib.
Add new job to CircleCI to build and archive the final library. This job should only run from master branch, however, currently have it enabled for all branches for testing.

* [amsdk-9772] Add CircleCI step to create and archive lib.
Splits CI config into three jobs, one for setting up the workspace, one for testing, and one for deploying the lib. The first job sets up the workspace and persists it for the other two downstream jobs.

* [amsdk-9772] Remove use of multiple jobs and add simply 'deploy' step.
Remove configurations for separate jobs which use a stored workspace and instead use 'deploy' step which checks the 'CIRCLE_BRANCH' env varable and if 'master' will build the library file.
Note, in this commit the 'deploy' step checks for branch 'pull/29' for testing.

* [amsdk-9772] CircleCI run 'deploy' step only on 'master' branch