[AMSDK-10131] Archive AEPCommerceDemoApp in CI (#35)
* [amsdk-10131] Adding new/blank AEPCommerceDemoApp
New AEPCommerceDemoApp.xcodeproj is under build/xcode and will be moved in later commits to demo folder once project files are setup.

* [amsdk-10131] Move AEPCommerceDemoApp files to new project.

* [amsdk-10131] Move xdmlib under new ACPCommerceDemoApp project folder.

* [amsdk-10131] update Podfile for new AEPCommerceDemoApp project

* [amsdk-10131] Adding storyboards back into commerce demo app project.

* [amsdk-10131] move new AEPCommerceDemoApp project from build/xcode to demo folder.

* [amsdk-10131] Add xdmlib to Podfile inorder to add ACPExperiencePlatform pod to target

* [amsdk-10131] Remove reference to the moved xdmlib and commerece demo app files.

* [amsdk-10131] Add AEPCommerceDemoApp Xcode project, workspace, and Podfile.
Creates a separate Xcode project for the AEPCommerceDemoApp which includes the ACPExperienePlatform as a library file. To use, copy the libACPExperiencePlatform.a file and ACPExperiencePlatform.swiftmodule folder to the libs/ folder. Open the AEPCommerceDemoApp.xcworkspace file and run the AEPCommerceDemoApp target.

* [amsdk-10131] Readd AEPCommerceDemoApp target and source files to ACPExperiencePlatform project.
The AEPCommerceDemoApp files are included here as source so it is convenient to build the app while developing the extension. However, as there is a separate Xcode project for the demo app, any file additions or deletions, or any project changes will also need to be done in the demo/AEPCommerceDemoApp/AEPCommerceDemoApp.xcworkspace as well.

* [amsdk-10131] Remove imports of xdmlib as it is no longer a library but instead built as source.

* [amsdk-10131] Copy swiftmodule files to out folder and suffix with extension version.

* [amsdk-10131] Update Makefile to generate zip archive of AEPCommerceDemoApp.

* [amsdk-10131] Add scheme in AEPCommerceDemoApp project to build demo app.

* [amsdk-10131] Add Makefile for AEPCommerceDemoApp.
Move make targets for archiving demo app from build/xcode/Makefile to new demo/AEPCommerceDemoApp/Makefile. Add targets to /Makefile to build and archive demo app.

* [amsdk-10131] Update AEPCommerceDemoApp project with pods

* [amsdk-10131] Add steps to build and archive demo app in CircleCI config

* [amsdk-10131] Build all and archive demo app for all branches for testing.
Remove the 'master' branch check in the deploy step for testing.

* [amsdk-10131] Build library before building demo app in CI config.

* [amsdk-10131] Add verbose flag to Codecov command in CI config

* [amsdk-10131] Bundle library and swiftmodule as zip file in out folder when building from Makefile.

* [amsdk-10131] Disable 's3' feature in Codecov uploader.
This is an attempt to fix the broken Codecov uploads. A forum post suggests this may fix the issue, https://community.codecov.io/t/upload-fails-repeatedly-on-bitrise/420/5.

* [amsdk-10131] Update pods to latest ACPCore (2.7.0) and ACPGriffon (1.1.2).

* [amsdk-10131] In copy-swiftmodule Makefile target, move swiftmodule directory and subdirectories.
In Swift 5.2.2 the compiler generates a Project folder under the swiftmodule folder. Using 'mv' to move the folders for each architecture will fail if it has subdirectories. Using tar to compress and extract the directories instead.

* [amsdk-10131] Enable Build_Library_For_Distribution flag in ACPExperiencePlatform lib target

* [amsdk-10131] Use Xcode 11.1 CircleCI container.
We would like to use the lowest version of Xcode possible to build the Swift library as the lib will only work when compiled for Swift versions >= the lib's Swift version. However, we'd also like to use a CocoaPods version with default support for CDN, which is v1.8.0+.

* [amsdk-10131] Fix CircleCI container vesion.
Remember to run 'circleci config validate' before pushing changes :)

* [amsdk-10131] Remove check and installation of Gemfile as CI is using default version of CocoaPods and not Gemfile version.

* [amsdk-10131] Rename class ACPExperiencePlatform to ExperiencePlatform

* [amsdk-10131] Rename Xcode project from ACPExperiencePlatform to AEPExperiencePlatform

* [amsdk-10131] Rename module in source to AEPExperiencePlatform

* [amsdk-10131] Remove ACPExperiencePlatform.xcworkspace

* [amsdk-10131] Update Makefiles with AEPExperiencePlatform name

* [amsdk-10131] Update AEPCommerceDemoApp to use libAEPExperiencePlatform.a

* [amsdk-10131] Update CircleCI config to reference new AEPExperiencePlatform name.

* [amsdk-10131] Update CircleCI config to build lib and demo app only on 'master' branch.